Translate default blocks and category names
We do not want to trust the quality of the translation of Blockly itself.
We have to prepare the translation ourselves.

For this purpose, this commit is to add blocks with the same name (id)
as the default/builtin blocks.

Do not think that the strings can be translated "word by word". Short words
like "do", "add", "if", "and" etc have multiple meanings and should be
translated into different words in different contexts. To solve this
issue, I have marked up these words as like "_repeat_do" and "_while_do".
Please be aware you have to add translations for these special strings
**even for English**.