Markup default variable name for translation
The default variable name was "item". We markup it as `lf("_default_var")`
instead of `lf("item")` because the word "item" is very short and general
purpose word. The word "item" should be translated into several different
words or phrases in different contexts.

For English, please add a translation from "_default_var" to "item".

For non-European languages, do not translate "item" literally because
only ascii alphabet or numbers can be used for variable names.