Add `{id:}` to short translation source strings
Short words may have multiple meanings, and such words should be
translated into different words or sentences in different contexts.
The `{id:...}` part in translation source strings will not appear in web
pages even when the locale is English or when no translation is available.
We can use `{id:...}` to identify the context.