Update to about.md to correct hyperlinks
The proposed change is for the micro:bit and Adafruit Circuit Playground Express! repository hyperlinks. They do not point to an repository, and have had the "microsoft" added into the user part of the link. the format for github repository links is "https://github.com/<username>/<repository>" with an example being "https://github.com/microsoft/pxt".