Fixed an issue where while using only the A, B button and Pin (but no thermometer or light level) in the microbit simulator, a space/padding was moving the simulator. (#2591)
After fixing this issue, the tag used as an aria-live was visible, so I hidden it too.
Finally, I fixed an issue where the simulator's UI was not correctly displayed at loading on FireFox while using only the A, B button and Pin (but no thermometer or light level). It had an impact on tabbing into the Pin.