Text encouraging use of VSCode.
Committed separately.

I immediately dove in using vim. In part, I had in the back of my mind
that VSCode was somehow going to be a heaviweight development
environment. (Studio came to mind.) However, the TypeScript integration
is excellent in VSCode, and that has helped.

The goal was to encourage those who come after me not to skip the idea
of using this editor in lieu of their particular favorite, at least not
initially.