Clarifying getting-started.md.
These changes are all of a piece; because I change some language
(consistently, I hope) throughout the page, I have not split these
changes into multiple commits.

I was tripped up by this page, because of the use of the word "package"
throughout. I did not understand, essentially, that the "package" is
just a "project" in which you edit main.ts. There is, apparently,
nothing particularly "special" about a package.

However, this meant that I was confused as to how to test, because
packages/projects looked the same. Hence, my changes attempt to make
this more explicit.

I also spent some time trying to figure out why my Bitbucket repository
would not work. I assumed that a git repository was a git repository. I
was wrong, apparently. So, I added text to clarify this as well.