Documentation updates to cli.md and getting-started.md (#3081)
* Clarifying text re: projects.

I found this text to be unclear when I was working my way through. The
"myworkspace" reference does not exist in the tree, and this language
about the "projects" directory is more in keeping with the repository
that is currently being checked out.

* Text encouraging use of VSCode.

Committed separately.

I immediately dove in using vim. In part, I had in the back of my mind
that VSCode was somehow going to be a heaviweight development
environment. (Studio came to mind.) However, the TypeScript integration
is excellent in VSCode, and that has helped.

The goal was to encourage those who come after me not to skip the idea
of using this editor in lieu of their particular favorite, at least not
initially.

* Clarifying getting-started.md.

These changes are all of a piece; because I change some language
(consistently, I hope) throughout the page, I have not split these
changes into multiple commits.

I was tripped up by this page, because of the use of the word "package"
throughout. I did not understand, essentially, that the "package" is
just a "project" in which you edit main.ts. There is, apparently,
nothing particularly "special" about a package.

However, this meant that I was confused as to how to test, because
packages/projects looked the same. Hence, my changes attempt to make
this more explicit.

I also spent some time trying to figure out why my Bitbucket repository
would not work. I assumed that a git repository was a git repository. I
was wrong, apparently. So, I added text to clarify this as well.

* Quotes, not asterisks.

I decided, for consistency, I should use quotes around the project name,
and not quote it in one place, and italicize it in another.