Adding ability to run getters in the debugger.
We add the ability to run getters in the debugger. When sending a
variable's children request, now we can specify which ones of the class
getters we should call to obtain values.

On the webapp side, we use this api to get getters marked to be
displayed in their comment attributes.