initial hack solution
fixes tabbing / screenreader focus, but is really gross.
This just bumps the tabindex of any focusable element to 'disabledtabindex' and
removes the elements from tab order, and then returns the tab index on exit.

Problematic things;
- would break on nested modals - easy enough to handle (suffix prop name with modal id?) but not great
- elements changing behind the modal would break things (not a thing in docs at the moment, but...)
- feels fragile

I wanna find a better solution but leaving in for now, have to deal with it before merging though