Remove the parent origin from the simulator search params
When we decide to check the origin of received messages in the simulator,
as well as when we post messages back to the simulator frame's parent window,
then we'll should put this back. That way, we'll be able to do something with
it, vs leaving it and having it do nothing.