Fix Asset Create Button Formatting
Specify flex properties for the common-button-flex contained within asset-editor-create-button, ensuring the content of the button (image & text) will wrap as needed to fit within the shape of the button.

Previously, these buttons relied on white-space wrapping to get this behavior, but that has since been disabled and the flex approach is hopefully more robust.