Fix Python Preferred Editor for Tutorials
When Python is set as the preferred editor for a tutorial, we would load blocks instead. This happened because the main.py file didn't exist at the time we were trying to load it.

To fix the issue, if the preferred file does not exist, we can create it rather than falling back on the blocks file.