Remove Width from Invisible Scrollbar CSS (#9016)
The custom width that was being applied is smaller than the actual scrollbar. As a result, when we showed/hid the scrollbar, the width of the simulator would change, and this oppened up an edge case where the change in width (and height) from adding/removing the scrollbar would triggers a change in whether or not the scrollbar is needed, so we'd end up looping, adding and removing the scrollbar constantly.

By removing the width from the invisible scrollbar CSS, the invisible scrollbar width simply defaults to the normal scrollbar width, so there's no size change when the bar is shown or hidden.

The comment on this line indicates that it was used to fix a "sim breakpoint" issue. We aren't certain what this issue was but I've toyed around with breakpoints and resizing the sim, and so far I haven't seen any unexpected behavior...