Fix Image Editor Top Bar Button Sizing
When a tutorial was open, there was a size threshold where the text in the image editor's top buttons would overrun the button bounds. This happened because the tutorial pane severely limited the horizontal space allowed for the popup, so we weren't yet in tablet mode when considering the whole available screen size, but the space allowed for the popup was too small.

To address this, I allow the popup to overlay the tutorial if the container's size is smaller than the tablet threshold (as opposed to the whole window). I also increased the weight on the buttons' section of the flexbox so they aren't quite so restricted. Combined, these give us enough flexibility to still look okay until we get to tablet mode, at which point the buttons change into their smaller form.

I considered trying to change the buttons form earlier (rather than modifying the popup overlay behavior), but that all happens in css using media screen size, and I didn't see a good way to change that without bringing all that into javascript, which seemed less ideal, especially since this affects a fairly narrow range of screen sizes and hopefully won't be hit too often.