Fix Image Editor Top Bar Button Sizing (#9054)
* Fix Image Editor Top Bar Button Sizing

When a tutorial is open, there's a size threshold where the text in the image editor's top buttons overruns the buttons' bounds. This happens because the tutorial pane severely limits the horizontal space allowed for the popup, but we aren't in tablet mode yet when considering the whole available screen size. Regardless, the space allowed for the popup is too small.

To address this, I've allowed the popup to overlay the tutorial if the container's size is smaller than the tablet threshold (as opposed to the whole window). I also increased the weight on the buttons' section of the flexbox so they aren't quite so restricted on either side. Combined, these give us enough flexibility to still look okay until we get to tablet mode, at which point the buttons change into their smaller form and we're good to go. (The tutorial pane is also moved at this point, and still covered up by the popup.)