When deciding autoRun, only consider emptyRunCode on first run.
The emptyRunCode var is a string with code in it, so !!emptyRunCode was always returning true (as long as the string is set, which it usually is). As a result, we were always returning true for autoRunOnStart(), even if light mode was enabled. As I understand it, we should only to an empty code run if it's also the first run, so I've added an additional check to account for that.