Disable AutoRun in Light Mode (#9072)
The emptyRunCode var is a string with code in it, so !!emptyRunCode always returns true (as long as the string is set in pxttarget.json, which it usually is). As a result, we always return true for autoRunOnStart(), even if light mode is enabled. As I understand it, we should only do an empty code run if it's also the first run, so I've added an additional check to account for that.

This fixes the first part of microsoft/pxt-arcade#2313: "Autorun should always be disabled for light mode"