Add a border radius to skill map learning objective tags (#9075)
Currently, the skillmap info panel has boxes that display a learning objectives, and they look very similar to the buttons lower down on the page, which can be confusing since the learning objective boxes are not interactive. This PR changes the shape of the learning objective containers so they're more visually distinct from the buttons.