Add Prefix to Tailwind CSS Classes (#9115)
This helps avoid conflicts with existing css classes, like "hidden", which can break tailwind behavior.

For example, I noticed an issue where tailwind could not properly adjust if something was shown or hidden based on screen size. The tailwind css classes to make that happen are hidden md:block. According to the tailwind docs, this should hide the item when the screen is smaller than md and show the item when the screen is md or larger. However, the behavior I saw was that the item was always hidden, regardless of screen size. This was happening because we already have a hidden css class in our common theme. That class prevented tailwind from being able to override the display value on md+ screen sizes.

We could do a one-off workaround for hidden, but that may be error-prone if devs aren't aware of it, and it's possible we may encounter similar issues for other classes later on, so adding a prefix seemed like a more robust solution. And better to do it now while it's still a manageable amount to change.

Ideally, this doesn't have any functional impact, though it's possible I missed the prefix in a few places in which case some layout may look off. We may need to just update those as we notice them, though I'm hoping there aren't many.

Tailwind docs on prefixes: https://tailwindcss.com/docs/configuration#prefix