Add Multiplayer Header Bar (#9117)
Add a header bar for multiplayer. It's loosely based on the header bar implementation in skillmaps, though I found it tricky to reuse much of it since we're using tailwind css and skillmap is not. Currently contains stubs for reporting abuse and getting help. Not sure how those will work, yet.