Create Single Game Window for Multiplayer Host and Join States (#9124)
The primary change here is to create a single "GamePage" to host the sim, players, and reaction content. The HostGame and JoinGame components have been repurposed/reduced into simple pages where a user can enter a game link or join code respectively. In the mid-to-long term, both HostGame and JoinGame should be entirely skippable with direct links (i.e. a "Host a Multiplayer Session for this Game" link would bypass HostGame, and a "Join Link" would bypass JoinGame. Both would go straight to the GamePage lobby state).

Meanwhile, the "lobby" (both join and host) is now represented with modals. They're not a perfect 1:1 with designs, but it's a start in that direction. Unclear if these will actually be modals in the long-term.

I've also done some initial UI work on the presence and reactions to bring them closer to the designs, but this isn't complete. It's still very rough around the edges, but this PR was already getting too big so I decided to send as-is, then we can finalize the UI piecemeal in later changes.