Update Multiplayer Lobby Appearance  (#9150)
This changes both host and join lobbies so they are non-modal and closer to the final designs. The sim exists and is loaded in the background, just hidden.

There are a few things I did not include in this change, because I think they'll work better as separate changes (and will take more time). Those are:
1. Creating and adding the full join link (https://github.com/microsoft/pxt-arcade/issues/5180).
2. QR Code (https://github.com/microsoft/pxt-arcade/issues/5181).
3. Left the "join" lobby text as-is. Designs have instructions on how to send emojis with the keyboard, but I omitted those because we don't have the keyboard controls yet. (https://github.com/microsoft/pxt-arcade/issues/5182)

Fixes https://github.com/microsoft/pxt-arcade/issues/5145
Fixes https://github.com/microsoft/pxt-arcade/issues/5146