Add Beta Indicator to Multiplayer Pages (#9157)
This adds a little "beta" tag to the home page, lobbies, and game page. Margins on the sim were a bit weird, so I just put it in the corner of the page in that scenario for now. If needed, we can revisit this later, but it may require css updates on the sim-side of things.