Fix Remix on Joined Games (#9177)
We weren't setting the game ID in state when a player joined a game, so we couldn't show the remix game button. With this change, it should now get saved in the GameInfo. I've also consolidated gameId variables, since we had one directly in state and one in state.gameState.