Add Functionality to the Multiplayer Keyboard Controls Button & Move Remix When Keyboard Controls is Hidden (#9196)
With this change, the keyboard controls button now displays a popup with the controls based on the existing popup we have for the code editor sim (and reusing several of the css classes there). The popup dismisses when the user clicks away.

The change also moves the "Remix" button to the right when the keyboard controls button is hidden on smaller screens.