Make Full Join Code Clickable for Copying Link (#9210)
This makes the full join code clickable to copy the join link, rather than just the icon to the side. To make it work in a way that felt smooth, I had to do some refactoring of the CopyButton to allow for a label attached to the button (rather than managing them separately). It will also display a toast when the copy is successful.