Cherry-Pick Downloading Multiple Projects as Zip to Stable7.0 (#9233)
Cherrypick 3656df4. I had to make a few changes so it could build on the older branch. Namely:

1. Use index instead of id to select headers.
2. Use sui.Modal instead of react-common Modal, since react-common Modal does not exist here.

This is needed so we can support the functionality in Microbit v4, which is what the old WinJS windows app uses. We're asking users to migrate to a new app, and this functionality allows them to transfer all of their projects over relatively easily.