Remove to do's that don't require action anymore.
Removed note to deprecate tutorialValidationRules. Going to wait and see how this approach goes, first.

Removed To Do about caching in a different location (so we have a new global validator instance for each step). We don't need that yet, and for now, this is more efficient. Can change later if needed.