Code Validation Infrastructure + Blocks Exist Validator (#9300)
This change adds a new infrastructure for code validation in tutorials, and includes one validator, which checks if the highlighted blocks in a tutorial step's answer key exist in the workspace, and that they're connected to the code.

Note that the new `MarkdownMetadataParser` file is all just cut/paste code from `skillMapParser`. I moved it to pxtlib so we can access it to parse the validation markdown, too.

Sample syntax for specifying a validator in tutorial markdown:
```validation.global
# BlocksExistValidator
```

You could also enable it for a single step using validation.local, much like blockConfig, and if you wanted to enable it globally but disable it for a single step, then you can enable it as above and add this in the local step you wish to disable it on:
```validation.local
# BlocksExistValidator
* enabled: false
```

If validation fails, a popup will appear, which does not close unless the user selects "Continue Anyway" or the close button. If the user clicks "Show Hint" it will update to display the answer key and some descriptive text. I'm still trying to figure out why there's a border around the blocks, but didn't want to block the check-in on it. Also, in the future, it would be best to show just the missing/detached blocks instead of the whole answer key, but that wasn't in scope for this MVP.