Remove caching of the tutorial validators.
When a tutorial is reloaded, these objects get serialized and then deserialized, but the "execute" method is not able to be stored in the serialized format, so it's lost on the cached object when the tutorial is loaded again, so when we try to call it, we get an error that it doesn't exist.

This cache isn't saving us from much work right now (validation construction is relatively cheap), so I've opted to simply remove it. If we decide we need it later on, we could cache it elsewhere (like on the TutorialContainer itself) and make sure it's cleared/refreshed as needed.