Fix Tutorial Next Button After Refresh (#9314)
We were seeing an issue with code validation, where after refreshing the page, the Next button in a tutorial no longer worked. To fix it, I've removed the caching of tutorial validators on the TutorialInfo and TutorialStepInfo objects.

When a tutorial is reloaded, these objects get serialized and then deserialized, but the "execute" method is not able to be stored in the serialized format, so it's lost in the process. As a result, when we try to call it, we get an error that it doesn't exist.

This cache isn't saving us from much work right now anyway (validation construction is relatively cheap), so I've opted to simply remove it. If we decide we need it later on, we could cache it elsewhere (like on the TutorialContainer itself) and make sure it's cleared/refreshed as needed.