Reset db version to 2, this way we don't trigger an update which can cause the tutorial load to hang if the db is being used by another tab when the request comes in to update it.
I think the "proper" fix would be to ensure we clean up our db connections so we don't hold onto it unnecessarily, but that would be a big change and isn't worth it for this edge case.

It should be safe to not update the version since we have null checks on the field anyway, so we (hopefully) won't crash if it's missing.