Fix Answer Key Causing Code Validation to Detect Error Incorrectly (#9346)
Not 100% consistent, but sometimes, if the user opens the answer key and then tries to proceed to the next step, validation will flag an error, even if the user has the correct code. This happens because, in this scenario, the call to Blockly.getMainWorkspace() sometimes doesn't return the user's actual workspace. Instead, it returns a workspace with no blocks in it.

To fix this, I've gotten rid of the call to Blockly.getMainWorkspace and instead grab it directly from the ProjectView editor (using an existing getBlocksEditor() function)