Add option to change language in Skillmap (#9382)
This adds an option to the skillmap gear menu to change the language. The UI is based off of the main webapp's UI, but it's been reimplemented in react-common. For now, I have not added the "Translate the Editor" button.

Eventually, it would be good to consolidate to a single language selector implementation, but I've just used it in skillmap for now. Once we've built confidence in it, we can go through and replace the other implementations with the react-common one. (Follow up task filed, see below.)

Links to follow-up items:
Show language option when an activity is open: https://github.com/microsoft/pxt-arcade/issues/5682
Unify language selectors: https://github.com/microsoft/pxt-arcade/issues/5684

Fixes https://github.com/microsoft/pxt-arcade/issues/5427