Make Tutorial Background Change with Green Screen Mode (#9400)
The new tutorial backgrounds weren't being set to transparent when in greenscreen mode.

When I made the switch, I also noticed green screen mode was adding a scroll bar that allowed the user to scroll the entire site horizontally. This was happening because the video element that gets created was larger than the site itself (it gets scaled up to fill the screen, but this usually means width becomes larger than 100% so height can be 100% while maintaining the correct aspect ratio). To avoid this behavior, I set 100% width in the video element itself rather than the css, then cap the width to 100% in css and set object-fit to crop the video element appropriately, so it doesn't shrink down again when css limits its size.

I also noticed that, since the tutorial pane extends 100% vertically on desktop and the editor tools pane (with the download button) is pushed to the right, it looked weird to have the green background on the tutorial pane but not the editor tools bar. In non-tutorial layouts, however, the editor tools bar keeps its white background, and I didn't want to change that. Instead, I decided to leave the editor tools bar at 100% width when green screen is enabled. I'm not sure that this is the best way to do it, but it's relatively cheap and I think a reasonable trade off. Could revisit if we get user feedback, or if any team members feel strongly that another approach would be preferable.