Sim-Sidebar Layout for Tutorials (#9430)
This adds an optional feature flag that allows us to switch tutorials to a layout with the sim on the left-hand side and instructions at the top. It's somewhat reminiscent of the older tutorial layout, mixed with the tablet/mobile layout in the current tutorial infrastructure. The changes only really affect desktop view. In mobile and tablet, the layout is relatively unchanged.