Make editor sidebar css selector specificity closer to how it was originally.
Changing from an id to a class made it less specific, so it was getting prioritized differently in the styling which caused a few bugs. This should help minimize the impact by adding the simulator id to all the rules where it's relevant. I did have to make a few adjustments to compensate for it.