Hide Tutorial Instructions in Fullscreen Mode and Fix CSS Specificity  (#9442)
This change ensures tutorial instructions are hidden in fullscreen sim mode.

It also goes through and adds #simulator to the places where I used .editor-sidebar so that the css specificity is closer to what it was before my changes in #9430. It used to be id-level and I changed it to class level, which had layout side-effects because the priority of certain styles was now different. The hope is that this will reduce the likelihood of those bugs.