Fix Minisim Display in Python/JS tutorials.
The check that displays the minisim when sidepanel recieves new props was causing a minisim to get displayed on javascript and python tutorials, when it should be getting shown on the sidebar instead. This was happening because the check was always showing the minisim anytime a tutorial was started, but that should actually only happen if we're not in the simSidebar layout.