Fix Minisim Display in Python/JS tutorials. (#9447)
The check that runs when sidepanel receives new props was causing the minisim to get displayed on javascript and python tutorials, but if we're in sim-sidebar layout, it should get shown on the sidebar instead. I simply forgot to update this check to account for the new layout, so I've added a condition to ensure it only shows the minisim if we're not in "sim-sidebar" mode.