More Tutorial Layout Fixes (#9452)
This addresses the following issues:
1. Increase tutorial height so it doesn't change as much.
2. Even out padding around tutorial controls.
3. Border under instructions in high contrast.

The best way I could find to then distribute the instruction content evenly throughout the space (once I'd given it a more-fixed height) was to switch the div containing the instructions (tutorial-content-bkg) from a block display to flex, which worked fairly well, but it also necessitated that I switch the immersive reader button to use absolute positioning, since it doesn't fit well into the flex layout.

While I was looking into adding the bottom border for instructions in high contrast mode, I realized the border that's supposed to be there normally was not getting applied either, because the css selector was looking for .topInstructionsWrapper > #simulator > .editor-sidebar but the actual hierarchy is #simulator > .topInstructionsWrapper > .editor-sidebar. In this case, #simulator is unnecessary anyway, so I just removed it.