Add Duration to Sound Effect Editor Duration Dropdown (#9521)
Add a dropdown menu to the sound effect editor's duration field. To do this, I added an optional options property that can be passed into a react-common Input. When present, a new button will appear that allows you to expand a dropdown list of options and select a value from there. Options are passed in as a map, where the text that gets displayed in the dropdown (i.e. "5 seconds") is mapped to the value that gets put in the Input when the option is selected (i.e. "5000").