Make Top-Instructions Resizable (#9533)
This makes the instructions resizable (when in the top-instructions view). I created a new VerticalResizeContainer so it will theoretically be easier to make similar changes easier in the future, if we decide to make other components resizable. (I did consider adding horizontal resizing too, but since it's not needed for this change, I decided to hold off. Can be added later if/when we need it.)