Error with 'groups'-array
The old documentation states, that the order of the 'groups'-function is determing in which order the groups will apeaer in the toolbox-flyout. But this is not true, it depends on which function, with which group, is defined first in the namespace. This can easily be seen in the playground (https://makecode.com/playground#basic-groups). According to the old documentation the 'Fruits'-group should be the first one in the toolbox-flyout, but this is wrong.