Improve Share Page Sandbox Display on Small Screens (#9570) (#9571)
This partially addresses the issues seen in microsoft/pxt-microbit#4828. While it's not a perfect fix, it's a step in the right direction, and it's at least usable now. I wanted to get this out in time for the hotfix, since I suspect some of the remaining improvements will take me longer to implement.

This change fixes the sim sizing and stops it from disappearing on smaller screens, and it simply removes the sim toolbar for now (which is currently offscreen anyway).