Fix instruction resize when hint is open (#9576) (#9578)
The issue is not actually specific to touch. When the hint is open, the popup is considered a child of the VerticalResizeContainer that contains the instructions, and it causes the resize "OnPointerDown" event to fire even when the mouse is below the main instruction container. Since I didn't add a max bounds check on before entering resize mode, it resize would trigger even when the mouse was below the instruction pane. We can stop this by checking that the mouse Y is less than the container's height before entering resize mode.