Make Minecraft Play Button Clickable (#9682)
Fixes microsoft/pxt-minecraft#2338.

The play button was being covered up by editor sidebar, which was invisible but still capturing mouse clicks. We can simply hide the editor sidebar in headless mode, since it's empty anyway, but it's still needed in tutorials to show the instructions, so I've moved some of the css into tutorial-sidebar.less instead.