Enable GitHub Release Zone in Main Branches. (#9742)
This is not a perfect fix. I tried going through and making main and master equivalent, but struggled with how to determine which one is used by a repo when on a non-main/master branch. Didn't think it was worth spending a ton of time on, so I paused that and focused on this quick fix instead. At least now, you can push releases and create branches from a Main branch.