Have Teacher Tool Request Iframe to Run Evaluation for Specified Share Link (#9821)
At a high level, this change does the following:
1. Connects the iframe to the share link that has been provided (except in localhost, because cross-domain localhost <-> makecode.com messaging doesn't work)
2. Saves project metadata and rubric info in app state
3. Adds a request to the iframe to get evaluation results. This request includes the rubric json as a parameter.

A few notable points:
- Logging Service
I added a logging service for convenient, consistent output and telemetry. There's LogError, which sends a tick event and does a console error output, LogInfo which does a tick event and non-error console output, and LogDebug, which sends console output but only if you're running the app locally.

- State in MakeCode Frame
I went back and forth on this, but in the end, in order to be consistent with our other component usage, I pushed the state for the project metadata into the MakecodeFrame. I can pull this out and have it passed in as props instead if desired, but I wasn't sure if there was all that much value in doing so at the moment.

- Fix for Promise Resolution in makecodeEditorService
Without the fix, when a message got queued up by Caller A but not sent immediately (editor not ready), then another caller has to come in and clear the queue once the editor is ready (Caller B). This was working okay, except it would assign the message's handler to the resolve function from the promise associated with Caller B's request, not Caller A's. As such, when a response comes back for the message, we'd end up resolving Caller B's promise instead of Caller A's, so Caller A would never return. To fix this, I assign the message metadata (including the handler) when we queue the message rather than when we dequeue it. This way, the handler will always get set to the resolve function for the promise returned to the caller that tried to send the message initially.

- Renamed loadProjectAsync to loadProjectMetadataAsync
Wanted to avoid confusion between loading the metadata and the project "loading" in the iframe.

- Placeholders
The EvaluationResult class and editor-side handler for the runeval request are just placeholders until we have actual evaluation running. I expect these to change in the near future.

- Try it Here
Upload target: https://arcade.makecode.com/app/e620e5c33cf3af44b631781f582f4ebcc34223c7-c7c3f7a5ad--eval