Add Validator Plans (#9835)
This adds validator plans and sends those to the editor for evaluation (rather than a rubric). It also adjusts evaluation logic slightly so that, instead of sending all requirements at once, we send one plan at a time. This allows us to get results for each check separately, so fast-running checks can finish and be displayed while long-running checks are still in progress.

Validation plans are loaded into state very similarly to how the catalog gets loaded, with shared and target-specific files as well as additional test files. The main change here is in how we handle test catalog/plans. I've now put them in a test folder instead of having files with -test suffixes, which matches how we do it for tutorials and skillmaps. It's also a little tidier in the directory.