Make the iframe ready reset an intentional action and only call it when we change the iframe url.
The ref can be updated anytime and doesn't necessarily mean the content of the iframe has changed or refreshed, so we can't simply unset the ready flag whenever that updates (the app has already sent its editorcontentloaded event, so it won't send another to unblock us). Instead, we need to preserve our ready flag unless the actual contents of the iframe changes.

An alternative approach would be to send a "ping" any time we try to send a message but we're not ready, and switch to ready if the ping is successful, but that feels like it may be unnecessary at this stage.